Separated options into their own modules
