+++ packaging/suse/portusctl/lib/options/README.md
@@ -1,0 +1,1 @@
+This directory contains options from the `setup` command.
